requirements.txt
numpy
pillow
colorama